Refactor massive builder files into submodules.
